extracted custom collections to a separate module
