+++ collection/build.gradle.kts
@@ -1,0 +1,7 @@
+dependencies {
+    implementation(project(":model"))
+}
+
+tasks.bootJar {
+    enabled = false
+}
