+++ collection/src/main/java/executor/service/collection/queue/proxy/ProxySourceQueueHandler.java
@@ -1,8 +1,8 @@
-package executor.service.queue.proxy;
+package executor.service.collection.queue.proxy;
 
 
+import executor.service.collection.queue.QueueHandler;
 import executor.service.model.ProxyConfigHolder;
-import executor.service.queue.QueueHandler;
 
 public interface ProxySourceQueueHandler extends QueueHandler<ProxyConfigHolder> {
 }
