+++ executor-service/build.gradle.kts
@@ -5,6 +5,7 @@ dependencies {
     implementation(project(":logger"))
     implementation(project(":model"))
     implementation(project(":webdriver"))
+    implementation(project(":collection"))
 }
