+++ executor-service/src/main/java/executor/service/config/BeanConfiguration.java
@@ -1,11 +1,6 @@
 package executor.service.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import executor.service.queue.ThreadSafeQueueHandler;
-import executor.service.queue.proxy.ProxySourceQueueHandler;
-import executor.service.queue.proxy.ProxySourceQueueHandlerImpl;
-import executor.service.queue.scenario.ScenarioSourceQueueHandler;
-import executor.service.queue.scenario.ScenarioSourceQueueHandlerImpl;
 import okhttp3.OkHttpClient;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -24,17 +19,6 @@ public class BeanConfiguration {
     @Bean
     public ObjectMapper objectMapper() {
         return new ObjectMapper();
-    }
-
-
-    @Bean
-    public ProxySourceQueueHandler proxyQueueHandler() {
-        return new ProxySourceQueueHandlerImpl(new ThreadSafeQueueHandler<>());
-    }
-
-    @Bean
-    public ScenarioSourceQueueHandler scenarioQueueHandler() {
-        return new ScenarioSourceQueueHandlerImpl(new ThreadSafeQueueHandler<>());
     }
 
 }
