+++ executor-service/src/main/java/executor/service/execution/ExecutionServiceImpl.java
@@ -3,7 +3,7 @@ package executor.service.execution;
 import executor.service.execution.scenario.ScenarioExecutor;
 
 import executor.service.model.Scenario;
-import executor.service.queue.scenario.ScenarioSourceQueueHandler;
+import executor.service.collection.queue.scenario.ScenarioSourceQueueHandler;
 import org.openqa.selenium.WebDriver;
 import org.springframework.stereotype.Service;
