+++ executor-service/src/test/java/executor/service/execution/ExecutionServiceImplTest.java
@@ -3,7 +3,7 @@ package executor.service.execution;
 import executor.service.execution.scenario.ScenarioExecutor;
 import executor.service.execution.scenario.ScenarioExecutorImpl;
 import executor.service.model.Scenario;
-import executor.service.queue.scenario.ScenarioSourceQueueHandler;
+import executor.service.collection.queue.scenario.ScenarioSourceQueueHandler;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.WebDriver;
