+++ executor-service/src/test/java/executor/service/execution/ParallelFlowExecutorImplTest.java
@@ -4,8 +4,8 @@ import executor.service.webdriver.factory.WebDriverProvider;
 import executor.service.execution.scenario.ScenarioExecutor;
 import executor.service.model.ProxyConfigHolder;
 import executor.service.model.ThreadPoolConfig;
-import executor.service.queue.proxy.ProxySourceQueueHandler;
-import executor.service.queue.scenario.ScenarioSourceQueueHandler;
+import executor.service.collection.queue.proxy.ProxySourceQueueHandler;
+import executor.service.collection.queue.scenario.ScenarioSourceQueueHandler;
 import executor.service.source.listener.LazyProxySourceListener;
 import executor.service.source.listener.ScenarioSourceListener;
 import executor.service.source.listener.SourceListener;
